Tests/LibWeb/Text/expected/sequential-navigation.txt

<BUTTON id="button">
<INPUT id="input">
<DIV id="tabindex-zero">
<BUTTON id="visibility-hidden-override">
